requirements.txt
berserk==0.10.0
boto==2.49.0
boto3==1.16.4
botocore==1.19.4
Brotli==1.0.9
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dash==1.16.2
dash-bootstrap-components==0.10.6
dash-core-components==1.12.1
dash-html-components==1.1.1
dash-renderer==1.8.2
dash-table==4.10.1
Deprecated==1.2.10
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
ndjson==0.3.1
numpy==1.19.2
pandas==1.1.2
plotly==4.11.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
retrying==1.3.3
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1